+++ include/avtp/acf/CanV2.h
@@ -364,8 +364,8 @@ OPEN1722_INLINE void Avtp_CanV2_SetCanIdentifier(Avtp_CanV2_t* msg, uint32_t can
  */
 OPEN1722_INLINE void Avtp_CanV2_SetPayloadLen(Avtp_CanV2_t* msg, uint16_t payloadLen) {
     uint16_t msgLenBytes = AVTP_CAN_V2_HEADER_LEN + payloadLen;
-    uint8_t pad = (4 - (msgLenBytes % 4)) % 4;
-    uint16_t msgLenQuadlets = (msgLenBytes + pad) / 4;
+    uint8_t pad = (uint8_t) (4 - (msgLenBytes % 4)) % 4;
+    uint16_t msgLenQuadlets = (uint16_t) ((msgLenBytes + pad) / 4);
     __Avtp_CanV2_SetField(AVTP_CAN_V2_FIELD_ACF_MSG_LENGTH, msgLenQuadlets);
     __Avtp_CanV2_SetField(AVTP_CAN_V2_FIELD_PAD, pad);
 }
