+++ include/avtp/acf/Gisf.h
@@ -435,8 +435,8 @@ OPEN1722_INLINE void Avtp_Gisf_SetLineNumber(Avtp_Gisf_t* msg, uint16_t lineNumb
  */
 OPEN1722_INLINE void Avtp_Gisf_SetPayloadLen(Avtp_Gisf_t* msg, uint16_t payloadLen) {
     uint16_t msgLenBytes = AVTP_GISF_HEADER_LEN + payloadLen;
-    uint8_t pad = (4 - (msgLenBytes % 4)) % 4;
-    uint16_t msgLenQuadlets = (msgLenBytes + pad) / 4;
+    uint8_t pad = (uint8_t) (4 - (msgLenBytes % 4)) % 4;
+    uint16_t msgLenQuadlets = (uint16_t) ((msgLenBytes + pad) / 4);
     __Avtp_Gisf_SetField(AVTP_GISF_FIELD_ACF_MSG_LENGTH, msgLenQuadlets);
     __Avtp_Gisf_SetField(AVTP_GISF_FIELD_PAD, pad);
 }
