+++ src/avtp/acf/custom/Vss.c
@@ -81,10 +81,11 @@ static uint16_t vss_read_clamped_length(const Avtp_Vss_t* pdu,
 
     if (offset + 2 + wire_len > declared) {
         /* Prefix fits but the payload it describes is too long → clamp. 
-         * Can never be negative, returns the maximum nuber of bytes that can 
-         * be read
-         */
-        return declared - (uint16_t)offset - 2;
+        * Clamp to the bytes available after the 2‑byte prefix.  The
+        * earlier guard ensures offset + 2 <= declared, so the result is
+        * non‑negative and fits uint16_t 
+        */
+        return (uint16_t) ((uint32_t) declared - offset - 2);
     }
     /* The normal case: All bytes declared fit in frame and can be read. */
     return wire_len;
